fix: ambiguous issue of param kind route
Route a [GET /users/:id]
Route b [GET /users/:uid/posts/:pid]

Request a [/users/1]
Request b [/users/1/posts/1]

The a will get the param map [id:1]
The b will get the param map [id:1 pid:1] (expected [uid:1 pid:1])

The a is good, but the b is wrong. This is because the param name
"uid" cannot cover the param name "id", so it appears as "id".