fix: multi routes cannot assign params properly
Route a [GET /users/:id]
Route b [GET /users/*]

Request a [/users/1]
Request b [/users/1/posts]

The a will get the param names [id] and the param values [1]
The b will get the param names [*] and the param values [1 1/posts]

The a is good, but the b is wrong. This is because the built-in func
append(), so the any kind param cannot cover the last param properly.