refactor: refine `Request` and `Response`
The request headers, cookies and params shouldn't always be parsed
automatically, but should be parsed when they are needed.

The response headers and cookies shouldn't be that hard to set up
(inspired by @SaulDoesCode).

Using slices instead of maps as much as possible. The slice is more
efficient than the map.